% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%>\% rhs
}
\value{
The result of applying rhs to lhs.
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
