% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot.tidylearn_model}
\alias{plot.tidylearn_model}
\title{Plot a tidylearn model}
\usage{
\method{plot}{tidylearn_model}(x, type = "diagnostics", ...)
}
\arguments{
\item{x}{A tidylearn model object}

\item{type}{Type of plot to create}

\item{...}{Additional arguments}
}
\value{
A ggplot object
}
\description{
Plot a tidylearn model
}
