% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot.tidysl_model}
\alias{plot.tidysl_model}
\title{Plot a tidysl model}
\usage{
\method{plot}{tidysl_model}(x, type = "diagnostics", ...)
}
\arguments{
\item{x}{A tidysl model object}

\item{type}{Type of plot to create}

\item{...}{Additional arguments}
}
\value{
A ggplot object
}
\description{
Plot a tidysl model
}
