% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{predict.tidylearn_json}
\alias{predict.tidylearn_json}
\title{Predict method for JSON models}
\usage{
\method{predict}{tidylearn_json}(object, new_data, type = "response", ...)
}
\arguments{
\item{object}{A tidylearn_json model object}

\item{new_data}{A data frame containing the new data}

\item{type}{Type of prediction}

\item{...}{Additional arguments}
}
\value{
Predictions
}
\description{
Predict method for JSON models
}
