% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{predict.tidylearn_onnx}
\alias{predict.tidylearn_onnx}
\title{Predict method for ONNX models}
\usage{
\method{predict}{tidylearn_onnx}(object, new_data, type = "response", ...)
}
\arguments{
\item{object}{A tidylearn_onnx model object}

\item{new_data}{A data frame containing the new data}

\item{type}{Type of prediction}

\item{...}{Additional arguments}
}
\value{
Predictions
}
\description{
Predict method for ONNX models
}
