% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{predict.tidysl_model}
\alias{predict.tidysl_model}
\title{Predict using a tidysl model}
\usage{
\method{predict}{tidysl_model}(object, new_data = NULL, type = "response", ...)
}
\arguments{
\item{object}{A tidysl model object}

\item{new_data}{A data frame containing the new data for prediction}

\item{type}{Type of prediction: "response", "prob", "class", etc.}

\item{...}{Additional arguments}
}
\value{
A tibble of predictions
}
\description{
Predict using a tidysl model
}
