% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{predict.tidysl_pmml}
\alias{predict.tidysl_pmml}
\title{Predict method for PMML models}
\usage{
\method{predict}{tidysl_pmml}(object, new_data, type = "response", ...)
}
\arguments{
\item{object}{A tidysl_pmml model object}

\item{new_data}{A data frame containing the new data}

\item{type}{Type of prediction}

\item{...}{Additional arguments}
}
\value{
Predictions
}
\description{
Predict method for PMML models
}
