% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.tidylearn_model}
\alias{print.tidylearn_model}
\title{Print method for tidylearn models}
\usage{
\method{print}{tidylearn_model}(x, ...)
}
\arguments{
\item{x}{A tidylearn model object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the model object
}
\description{
Print method for tidylearn models
}
