% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.tidysl_model}
\alias{print.tidysl_model}
\title{Print method for tidysl models}
\usage{
\method{print}{tidysl_model}(x, ...)
}
\arguments{
\item{x}{A tidysl model object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the model object
}
\description{
Print method for tidysl models
}
