% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.tidysl_pipeline}
\alias{print.tidysl_pipeline}
\title{Print a tidysl pipeline}
\usage{
\method{print}{tidysl_pipeline}(x, ...)
}
\arguments{
\item{x}{A tidysl pipeline object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the pipeline
}
\description{
Print a tidysl pipeline
}
