% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{summary.tidylearn_model}
\alias{summary.tidylearn_model}
\title{Summary method for tidylearn models}
\usage{
\method{summary}{tidylearn_model}(object, ...)
}
\arguments{
\item{object}{A tidylearn model object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the model summary
}
\description{
Summary method for tidylearn models
}
