% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{summary.tidysl_model}
\alias{summary.tidysl_model}
\title{Summary method for tidysl models}
\usage{
\method{summary}{tidysl_model}(object, ...)
}
\arguments{
\item{object}{A tidysl model object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the model summary
}
\description{
Summary method for tidysl models
}
