% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{summary.tidysl_pipeline}
\alias{summary.tidysl_pipeline}
\title{Summarize a tidysl pipeline}
\usage{
\method{summary}{tidysl_pipeline}(object, ...)
}
\arguments{
\item{object}{A tidysl pipeline object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the pipeline
}
\description{
Summarize a tidysl pipeline
}
