% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{tl_calc_regression_metrics}
\alias{tl_calc_regression_metrics}
\title{Calculate regression metrics}
\usage{
tl_calc_regression_metrics(
  actuals,
  predicted,
  metrics = c("rmse", "mae", "rsq", "mape"),
  ...
)
}
\arguments{
\item{actuals}{Actual values (ground truth)}

\item{predicted}{Predicted values}

\item{metrics}{Character vector of metrics to compute}

\item{...}{Additional arguments}
}
\value{
A tibble of evaluation metrics
}
\description{
Calculate regression metrics
}
\keyword{internal}
