% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tl_check_packages}
\alias{tl_check_packages}
\title{Helper function to check if required packages are installed}
\usage{
tl_check_packages(pkg, error = TRUE)
}
\arguments{
\item{pkg}{Character vector of required packages}

\item{error}{Logical; whether to throw an error if packages are missing}
}
\value{
Logical vector indicating if packages are installed
}
\description{
Helper function to check if required packages are installed
}
