% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tl_format_numbers}
\alias{tl_format_numbers}
\title{Format numbers with standardized precision}
\usage{
tl_format_numbers(x, digits = 3, format = "standard", zero_pad = FALSE)
}
\arguments{
\item{x}{Vector of numbers to format}

\item{digits}{Number of digits to round to}

\item{format}{Format to use: "standard", "scientific", "percent"}

\item{zero_pad}{Logical; whether to pad with zeros}
}
\value{
Character vector with formatted numbers
}
\description{
Format numbers with standardized precision
}
