% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{tl_from_caret}
\alias{tl_from_caret}
\title{Convert a caret model to a tidylearn model}
\usage{
tl_from_caret(model)
}
\arguments{
\item{model}{A caret model object}
}
\value{
A tidylearn model object
}
\description{
Convert a caret model to a tidylearn model
}
