% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tl_model}
\alias{tl_model}
\title{Create a tidysl model}
\usage{
tl_model(data, formula, method = "linear", ...)
}
\arguments{
\item{data}{A data frame containing the training data}

\item{formula}{A formula specifying the model}

\item{method}{The modeling method to use (e.g., "linear", "logistic", "tree", etc.)}

\item{...}{Additional arguments to pass to the underlying model function}
}
\value{
A tidysl model object
}
\description{
Create a tidysl model
}
