% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tl_partition_data}
\alias{tl_partition_data}
\title{Create a data partition for training and testing}
\usage{
tl_partition_data(data, prop = 0.75, strata = NULL, seed = NULL)
}
\arguments{
\item{data}{A data frame containing the data}

\item{prop}{Proportion of data to use for training}

\item{strata}{Optional variable name to use for stratified sampling}

\item{seed}{Random seed for reproducibility}
}
\value{
A list with training and testing indices
}
\description{
Create a data partition for training and testing
}
