% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\name{tl_plot_regularization_cv}
\alias{tl_plot_regularization_cv}
\title{Plot cross-validation results for a regularized regression model}
\usage{
tl_plot_regularization_cv(model, ...)
}
\arguments{
\item{model}{A tidysl regularized model object}

\item{...}{Additional arguments}
}
\value{
A ggplot object
}
\description{
Plot cross-validation results for a regularized regression model
}
