% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tl_scale_to_range}
\alias{tl_scale_to_range}
\title{Scale variables to a specified range}
\usage{
tl_scale_to_range(x, lower = 0, upper = 1, na.rm = TRUE)
}
\arguments{
\item{x}{Vector to scale}

\item{lower}{Lower bound of range}

\item{upper}{Upper bound of range}

\item{na.rm}{Logical; whether to remove NA values}
}
\value{
Scaled vector
}
\description{
Scale variables to a specified range
}
