% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{tl_to_tidymodels}
\alias{tl_to_tidymodels}
\title{Convert a tidylearn model to a tidymodels workflow}
\usage{
tl_to_tidymodels(model)
}
\arguments{
\item{model}{A tidylearn model object}
}
\value{
A tidymodels workflow object
}
\description{
Convert a tidylearn model to a tidymodels workflow
}
