% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{tl_from_tidymodels}
\alias{tl_from_tidymodels}
\title{Convert a tidymodels workflow to a tidysl model}
\usage{
tl_from_tidymodels(workflow)
}
\arguments{
\item{workflow}{A tidymodels workflow object}
}
\value{
A tidysl model object
}
\description{
Convert a tidymodels workflow to a tidysl model
}
